// src/app/interfaces/ICompetenciaSobreviviente.ts
export interface ICompetenciaSupervivencia{
    sobrevivienteID: number
    filtrado: number
    purificacion: number
    trabajoEquipo: number
    estructuraRefugio: number
    proteccionElementos: number
    seguridadEfectividad: number
    ejecucionMetodo: number
    orientacionPrecisa: number
    explicacionVariacion: number
    encenderFuego: number
    mantenerlo: number
    seguridadFuego: number
    seguridadHigiene: number
    comidaCocida: number
    creatividadCoccion:number
    bonus: number
    comentarios: string
    participanteID: number
}